+++ CHANGELOG.md
@@ -362,4 +362,6 @@
 ### Bugfixes
 
 - [Issue #398](https://github.com/influxdb/influxdb/issues/398). Support now() and NOW() in the query lang
+- [Issue #403](https://github.com/influxdb/influxdb/issues/403). Filtering should work with join queries
+- [Issue #404](https://github.com/influxdb/influxdb/issues/404). Filtering with invalid condition shouldn't crash the server
 - Close leveldb databases properly if we couldn't create a new Shard. See leveldb\_shard\_datastore\_test:131